+++ src/main/java/org/sonar/plugins/delphi/pmd/rules/ForbiddenIdentifierRule.java
@@ -1,6 +1,5 @@
 package org.sonar.plugins.delphi.pmd.rules;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
@@ -13,7 +12,7 @@ public class ForbiddenIdentifierRule extends AbstractDelphiRule {
   public static final PropertyDescriptor<List<String>> BLACKLISTED_NAMES =
       PropertyFactory.stringListProperty("blacklist")
           .desc("The list of forbidden identifiers. (case-insensitive)")
-          .defaultValue(Collections.emptyList())
+          .emptyDefaultValue()
           .build();
 
   private static final PropertyDescriptor<String> MESSAGE =